ChartScanAI_Shiny/requirements.txt
yfinance>=0.2.28
mplfinance>=0.12.10
pandas>=2.0.0
ultralytics>=8.0.0
azure-storage-blob>=12.19.0
python-dotenv>=1.0.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0